==> scripts/r/production/00_libraries.R <==
# Load libraries --------------------------------------------------------------
#
# Last update: 2022-03-16
#

library("purrr")
library("dplyr")
library("tidyr")
library("readr")
library("forcats")
library("glue")
library("here")
library("broom")
library("stringr")

library("brms")
library("bayestestR")
library("modelr")
library("emmeans")

library("ggplot2")
library("ggbeeswarm")
library("tidybayes")
library("bayesplot")
library("ggstance")
library("ggExtra")

library("future")
library("flextable")
library("academicWriteR")

# -----------------------------------------------------------------------------
